Introduce FrameStackInitializatioNode
- Try primitive before initializing a frame
- Refactor HandlePrimitiveFailedNode

+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/exceptions/PrimitiveExceptions.java
@@ -9,13 +9,13 @@ public final class PrimitiveExceptions {
 
     protected static class AbstractPrimitiveFailed extends ControlFlowException {
         private static final long serialVersionUID = 1L;
-        private final long reasonCode;
+        private final int reasonCode;
 
-        protected AbstractPrimitiveFailed(final long reasonCode) {
+        protected AbstractPrimitiveFailed(final int reasonCode) {
             this.reasonCode = reasonCode;
         }
 
-        public final long getReasonCode() {
+        public final int getReasonCode() {
             return reasonCode;
         }
     }
@@ -30,7 +30,7 @@ public final class PrimitiveExceptions {
         public static final PrimitiveFailed INAPPROPRIATE_OPERATION = new PrimitiveFailed(ERROR_TABLE.INAPPROPRIATE_OPERATION);
         public static final PrimitiveFailed INSUFFICIENT_OBJECT_MEMORY = new PrimitiveFailed(ERROR_TABLE.INSUFFICIENT_OBJECT_MEMORY);
 
-        private PrimitiveFailed(final long reasonCode) {
+        private PrimitiveFailed(final int reasonCode) {
             super(reasonCode);
         }
 
@@ -39,7 +39,7 @@ public final class PrimitiveExceptions {
             throw GENERIC_ERROR;
         }
 
-        public static PrimitiveFailed andTransferToInterpreter(final long reason) {
+        public static PrimitiveFailed andTransferToInterpreter(final int reason) {
             CompilerDirectives.transferToInterpreter();
             throw new PrimitiveFailed(reason);
         }
@@ -48,7 +48,7 @@ public final class PrimitiveExceptions {
     public static final class SimulationPrimitiveFailed extends AbstractPrimitiveFailed {
         private static final long serialVersionUID = 1L;
 
-        public SimulationPrimitiveFailed(final long reasonCode) {
+        public SimulationPrimitiveFailed(final int reasonCode) {
             super(reasonCode);
         }
     }
